Extract sync orchestration into a useSyncManager hook
App.tsx declared all nine useSync instances and owned the cross-ritual
orchestration inline: mutual reset (clearOthers), every per-ritual handler, the
seven-step sequential "Wielki Rytuał" full sync, the duplicates formatter, and
the aggregate result state. That is data/flow logic living in a component,
against the guidelines' Logic-Isolation rule.

Move it all into src/hooks/useSyncManager.ts, which returns the ritual
instances, the syncs array, anyError/isAnySyncLoading, the full-sync-results
state with a clear() helper, and every handler. App.tsx now just destructures
the hook and renders — it drops ~130 lines (409 → 282) and no longer imports
useSync or PREDEFINED_AWARDS. Behavior is unchanged.

Guidelines §2 documents useSyncManager as the orchestration hook.
